@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React, { useEffect, useState } from 'react';
 import './input.scss';
 
 interface Props {
@@ -6,45 +6,33 @@ interface Props {
   id: string;
   name: string;
   search: string;
-  list?: string;
   searchValue: (value: string) => void;
 }
 
-interface State {
-  inputValue: string;
-}
+function Input(props: Props): React.ReactNode {
+  const [inputValue, setInputValue] = useState('');
+  const { type, id, name, search, searchValue } = props;
 
-class Input extends Component<Props, State> {
-  constructor(props: Props) {
-    super(props);
-    this.state = {
-      inputValue: '',
-    };
-  }
+  useEffect(() => {
+    setInputValue(search);
+  }, [search]);
 
-  componentDidMount() {
-    this.setState({ inputValue: this.props.search });
-  }
-
-  onChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+  function onChange(event: React.ChangeEvent<HTMLInputElement>) {
     const { target } = event;
-    this.setState({ inputValue: target.value });
-    this.props.searchValue(target.value);
-  };
-
-  render() {
-    const { type, name, id } = this.props;
-    return (
-      <input
-        className={'input'}
-        type={type}
-        name={name}
-        id={id}
-        value={this.state.inputValue}
-        onChange={this.onChange}
-      />
-    );
+    setInputValue(target.value);
+    searchValue(target.value);
   }
+
+  return (
+    <input
+      className={'input'}
+      type={type}
+      name={name}
+      id={id}
+      value={inputValue}
+      onChange={onChange}
+    />
+  );
 }
 
 export default Input;
