@@ -1,4 +1,4 @@
-import { Component } from 'react';
+import React from 'react';
 import { Character, Episode, Location } from '../../interfaces.ts';
 import './result-item.scss';
 import ResultItemCharacter from './result-item-character.tsx';
@@ -9,18 +9,15 @@ interface Props {
   result: Character | Location | Episode;
 }
 
-type State = object;
+function ResultItem(props: Props): React.ReactNode {
+  const { result } = props;
 
-class ResultItem extends Component<Props, State> {
-  render() {
-    const { result } = this.props;
-    if ('image' in result && result.image)
-      return <ResultItemCharacter result={result} />;
-    if ('dimension' in result && result.dimension)
-      return <ResultItemLocation result={result} />;
-    if ('air_date' in result && result.air_date)
-      return <ResultItemEpisode result={result} />;
-  }
+  if ('image' in result && result.image)
+    return <ResultItemCharacter result={result} />;
+  if ('dimension' in result && result.dimension)
+    return <ResultItemLocation result={result} />;
+  if ('air_date' in result && result.air_date)
+    return <ResultItemEpisode result={result} />;
 }
 
 export default ResultItem;
