@@ -3,6 +3,9 @@ import { createRoot } from 'react-dom/client';
 import './index.css';
 import App from './App.tsx';
 import ErrorBoundary from './error-boundary.tsx';
+import { BrowserRouter, Navigate, Route, Routes } from 'react-router';
+import DetailedItem from './components/detailed-item/detailed-item.tsx';
+import NotFoundPage from './pages/NotFound/not-found.tsx';
 
 const root = document.getElementById('root');
 
@@ -10,7 +13,16 @@ if (root)
   createRoot(root).render(
     <StrictMode>
       <ErrorBoundary>
-        <App />
+        <BrowserRouter>
+          <Routes>
+            <Route path={'/'} element={<Navigate to={'search'} />} />
+            <Route path={'/search'} element={<App />} />
+            <Route path={'/search/:resource'} element={<App />}>
+              <Route index path={':id'} element={<DetailedItem />} />
+            </Route>
+            <Route path={'*'} element={<NotFoundPage />} />
+          </Routes>
+        </BrowserRouter>
       </ErrorBoundary>
     </StrictMode>
   );
